Add index.html
Previous commit history was a mess. This was fixed with liberal use of
git reset and git rebase, but unfortunately that flogged index.html
out of existence. Thus, included in this commit is a surviving copy of
index.html, preserved through luck and a stubborn text editor.

PREVIOUS CHANGES (chronological order)

commit 6876322d48b0ef0fce25f2e9e4a0f59cea63263f
Author: tyxchen <tyxchen@outlook.com>
Date:   Mon Feb 29 22:54:51 2016 -0500

    Integrate formatter into index.html

commit e429c0b2c822b795bad8072385043c3c3af8950e
Author: tyxchen <tyxchen@outlook.com>
Date:   Tue Mar 1 11:02:28 2016 -0500

    Add rudimentary styling, options panel

commit d72aa7a21826132b6cb4489d8cded9664acbf4e8
Author: tyxchen <tyxchen@outlook.com>
Date:   Tue Mar 1 13:39:02 2016 -0500

    Add auto-disabling of checkboxes and dynamic flags

commit 66163df6b0f6fc36e001899347ca9ab123044958
Author: tyxchen <tyxchen@outlook.com>
Date:   Tue Mar 1 15:24:24 2016 -0500

    Add dynamic loading of formatters